Added some working timer code. ~65ms is usual build time for current arena_locker level.

+++ Wasteland_OGRE/Code/Arena_Locker.cpp
@@ -57,7 +57,7 @@ void ArenaLocker::Setup(InputManager* Input,GraphicsManager* Graphics,GUIManager
 	testParams.setAgentRadius(.2f);
 	
 	RecastInterface recast(_scene,testParams);
-	recast.getRecastConfig().walkableRadius = .2f;
+	recast.getRecastConfig().walkableRadius = static_cast<int>(.2f);
 	recast.buildNavMesh(&testGeom);
 	recast.exportPolygonMeshToObj("RECAST_NAVMESH_TEST.obj");
 	recast.recastClean();
